Started Query to Git
